# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: BallotWave - Digital Voting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /elections:
    get:
      operationId: listElections
      tags: [elections]
      summary: List elections
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed, cancelled]
      responses:
        "200":
          description: List of elections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Election"
    post:
      operationId: createElection
      tags: [elections]
      summary: Create a new election
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateElectionRequest"
      responses:
        "201":
          description: Election created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{electionId}:
    get:
      operationId: getElection
      tags: [elections]
      summary: Get election by ID
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionDetail"
    patch:
      operationId: updateElection
      tags: [elections]
      summary: Update election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateElectionRequest"
      responses:
        "200":
          description: Election updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{electionId}/start:
    post:
      operationId: startElection
      tags: [elections]
      summary: Start an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{electionId}/close:
    post:
      operationId: closeElection
      tags: [elections]
      summary: Close an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Election"
  /elections/{electionId}/results:
    get:
      operationId: getElectionResults
      tags: [elections]
      summary: Get election results
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionResults"

  # Candidates
...
  /elections/{electionId}/vote:
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote in an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CastVoteRequest"
      responses:
        "200":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResponse"
        "400":
          description: Already voted or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /elections/{electionId}/check-voted:
    get:
      operationId: checkVoted
      tags: [votes]
      summary: Check if voter has voted
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vote status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteStatusResponse"

  # Payments
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate payment for voting access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentRequest"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
  /analytics/elections/{electionId}:
    get:
      operationId: getElectionAnalytics
      tags: [analytics]
      summary: Get detailed analytics for an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionAnalytics"

  # Users
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    # Auth Schemas
...
    Election:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        schoolId:
          type: string
        schoolName:
          type: string
        status:
          type: string
          enum: [draft, active, closed, cancelled]
        votingType:
          type: string
          enum: [web, ussd, both]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        requiresPayment:
          type: boolean
        votingFee:
          type: number
        currency:
          type: string
        totalVotes:
          type: integer
        totalCandidates:
          type: integer
        allowMultiplePositions:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, schoolId, status, votingType, startDate, endDate, requiresPayment, totalVotes, totalCandidates, createdAt]
    ElectionDetail:
      allOf:
        - $ref: "#/components/schemas/Election"
        - type: object
          properties:
            candidates:
              type: array
              items:
                $ref: "#/components/schemas/Candidate"
            positions:
              type: array
              items:
                type: string
    CreateElectionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        schoolId:
          type: string
        votingType:
          type: string
          enum: [web, ussd, both]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        requiresPayment:
          type: boolean
        votingFee:
          type: number
        currency:
          type: string
        allowMultiplePositions:
          type: boolean
      required: [title, schoolId, votingType, startDate, endDate]
    UpdateElectionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, closed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        requiresPayment:
          type: boolean
        votingFee:
          type: number
    ElectionResults:
      type: object
      properties:
        electionId:
          type: string
        title:
          type: string
        status:
          type: string
        totalVotes:
          type: integer
        turnoutPercentage:
          type: number
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionResult"
      required: [electionId, title, status, totalVotes, turnoutPercentage, positions]
    PositionResult:
      type: object
      properties:
        position:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/CandidateResult"
        winner:
          $ref: "#/components/schemas/CandidateResult"
      required: [position, candidates]
    CandidateResult:
      type: object
      properties:
        candidateId:
          type: string
        name:
          type: string
        voteCount:
          type: integer
        percentage:
          type: number
        photoUrl:
          type: string
      required: [candidateId, name, voteCount, percentage]

    # Candidates
    Candidate:
      type: object
      properties:
        id:
          type: string
        electionId:
          type: string
        userId:
          type: string
        name:
          type: string
        position:
          type: string
        manifesto:
          type: string
        photoUrl:
          type: string
        studentId:
          type: string
        department:
          type: string
        voteCount:
          type: integer
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, electionId, name, position, voteCount, isApproved, createdAt]
...
    CastVoteRequest:
      type: object
      properties:
        votes:
          type: array
          items:
            $ref: "#/components/schemas/VoteEntry"
        paymentReference:
          type: string
      required: [votes]
    VoteEntry:
      type: object
      properties:
        position:
          type: string
        candidateId:
          type: string
      required: [position, candidateId]
    VoteResponse:
      type: object
      properties:
        success:
          type: boolean
        receiptCode:
          type: string
        message:
          type: string
      required: [success, receiptCode, message]
    VoteStatusResponse:
      type: object
      properties:
        hasVoted:
          type: boolean
        receiptCode:
          type: string
      required: [hasVoted]

    # Payments
    InitiatePaymentRequest:
      type: object
      properties:
        electionId:
          type: string
        email:
          type: string
        phone:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [card, mobile_money, bank_transfer]
        mobileMoneyProvider:
          type: string
          enum: [mtn, vodafone, airteltigo]
      required: [electionId, email, amount, paymentMethod]
    PaymentInitResponse:
      type: object
      properties:
        reference:
          type: string
        authorizationUrl:
          type: string
        status:
          type: string
      required: [reference, status]
    VerifyPaymentRequest:
      type: object
      properties:
        reference:
          type: string
      required: [reference]
    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        reference:
          type: string
        status:
          type: string
        amount:
          type: number
        paidAt:
          type: string
          format: date-time
      required: [success, reference, status]

    # Analytics
...
    ElectionAnalytics:
      type: object
      properties:
        electionId:
          type: string
        title:
          type: string
        totalRegisteredVoters:
          type: integer
        totalVotes:
          type: integer
        turnoutPercentage:
          type: number
        votesByHour:
          type: array
          items:
            $ref: "#/components/schemas/HourCount"
        votesByMethod:
          type: array
          items:
            $ref: "#/components/schemas/MethodCount"
        fraudAlerts:
          type: integer
      required: [electionId, title, totalRegisteredVoters, totalVotes, turnoutPercentage, votesByHour, votesByMethod, fraudAlerts]
...
    HourCount:
      type: object
      properties:
        hour:
          type: string
        count:
          type: integer
      required: [hour, count]
    MethodCount:
      type: object
      properties:
        method:
          type: string
        count:
          type: integer
      required: [method, count]