# lib/api-zod/src/generated/types/initiatePaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BallotWave - Digital Voting Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { InitiatePaymentRequestMobileMoneyProvider } from "./initiatePaymentRequestMobileMoneyProvider";
import type { InitiatePaymentRequestPaymentMethod } from "./initiatePaymentRequestPaymentMethod";

export interface InitiatePaymentRequest {
  electionId: string;
  email: string;
  phone?: string;
  amount: number;
  paymentMethod: InitiatePaymentRequestPaymentMethod;
  mobileMoneyProvider?: InitiatePaymentRequestMobileMoneyProvider;
}