# lib/api-zod/src/generated/types/voteEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BallotWave - Digital Voting Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface VoteEntry {
  position: string;
  candidateId: string;
}